Update drf-oidc-auth to 0.10, fixing a bug breaking Keycloak compatibility
drf-oidc-auth had a bug whereby it tried to compare 'azp' claim against
allowed audiences. This is wrong as azp identifies they authorized
party, not the resource provider who is identified by audience.

Also django-helusers has a bug with the settings passed to drf-oidc-auth,
but this masks it for time being.